Renamed arrow icons
