+++ src/components/atoms/Svg/Icons/ArrowRight.jsx
