+++ src/components/atoms/Svg/index.jsx
@@ -3,10 +3,10 @@ export { default as MinusIcon } from "./Icons/Minus"
 export { default as CompareArrowIcon } from "./Icons/CompareArrow"
 export { default as EditIcon } from "./Icons/Edit"
 export { default as DeleteIcon } from "./Icons/Delete"
-export { default as DownIcon } from "./Icons/Down"
-export { default as RightIcon } from "./Icons/Right"
-export { default as LeftIcon } from "./Icons/Left"
-export { default as UpIcon } from "./Icons/Up"
+export { default as ArrowDownIcon } from "./Icons/ArrowDown"
+export { default as ArrowRightIcon } from "./Icons/ArrowRight"
+export { default as ArrowLeftIcon } from "./Icons/ArrowLeft"
+export { default as ArrowUpIcon } from "./Icons/ArrowUp"
 export { default as DiscordIcon } from "./Icons/Discord"
 export { default as TelegramIcon } from "./Icons/Telegram"
 export { default as GithubIcon } from "./Icons/Github"
